fix(ts/analyzer): Fix the evaluation order of a module declaration (#416)
**Description:**

While generating the dependency graph, `module Foo` used to be declared both as a type and a variable. This causes the module to be evaluated before the var-declaration, therefore making `foo` invisible to the module. However, it is [not currently valid to refer to modules in type annotations](https://www.typescriptlang.org/play?#code/LYewJgrgNgpgBAMRCOBvAUHLcYA8AOIATgC5wkCe+8SIAyiUQJYB2A5nALxwDOjrbdAF90ANwCGROADNkALkTI4AemVwWIMhKhMwYyTL7N2C2gDpaDYx1VwmJHjgLEHcSTAA0cABYgA7jCiMERAA), so they shouldn't be declared as one.