refactor(ts/env): Remove swc globals from `Env` (#473)
**Description:**

This PR removes `swc_globals` from `Env`. `Env` now does not manage `Marks`, so the user is responsible for providing the correct `Marks`. i.e. Created from the current `swc_common::GLOBALS`.

This change is made because the previous design was too error-prone. I found lots of mistakes due to `Env` in tests and fixed all of them.

Additionally, this PR exposes the top-level context to the `Analyzer`. The analyzer requires top-level context because of JSX.