chore: update debugger config (#729)
<!-- Note: CI script will automatically rebase your PR so please do not
rebase unless required -->

<!-- Thanks for submitting a pull request! Please provide enough
information so that others can review your pull request. -->

**Description:**
update debug file to run tests with debugger under the same condition as
conformance tests.
<!-- Explain the **motivation** for making this change. What existing
problem does the pull request solve? -->

**BREAKING CHANGE:**

<!--
If this PR introduces a breaking change, it must contain a notice for it
to be included in the CHANGELOG. Add description or remove entirely if
not breaking.

You may need to update `MIGRATION.md` for the breaking changes.
-->

**Related issue (if exists):**

Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>